Moved rotary encoder function to daskeyboard4.c
