+++ keyboards/handwired/daskeyboard/daskeyboard4/keymaps/default/keymap.c
@@ -43,14 +43,3 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
         KC_TRNS,  KC_TRNS,    KC_TRNS,                            KC_TRNS,                            KC_TRNS,    KC_TRNS,    KC_TRNS,    KC_TRNS,                      KC_TRNS,   KC_TRNS,  KC_TRNS,    KC_TRNS,              KC_TRNS
     )
 };
-
-// rotary encoder
-bool encoder_update_user(uint8_t index, bool clockwise) {
-    // Volume control
-    if (clockwise) {
-        tap_code(KC_VOLD);
-    } else {
-        tap_code(KC_VOLU);
-    }
-    return false;
-};
